fixed the moving config.json oops

+++ RPBot-Core/Program.cs
@@ -31,17 +31,28 @@ namespace RPBot
             RPClass.LoadData();
             var cfg = new Config();
             var json = string.Empty;
-            if (!File.Exists("../../config.json"))
+            if (!File.Exists("config.json"))
             {
-                json = JsonConvert.SerializeObject(cfg);
-                File.WriteAllText("../../config.json", json, new UTF8Encoding(false));
-                Console.WriteLine("Config file was not found, a new one was generated. Fill it with proper values and rerun this program");
-                Console.ReadKey();
+                if (!File.Exists("../../config.json"))
+                {
+                    json = JsonConvert.SerializeObject(cfg);
+                    File.WriteAllText("../../config.json", json, new UTF8Encoding(false));
+                    Console.WriteLine("Config file was not found, a new one was generated. Fill it with proper values and rerun this program");
+                    Console.ReadKey();
 
-                return;
+                    return;
+                }
+                else
+                {
+                    json = File.ReadAllText("../../config.json", new UTF8Encoding(false));
+                }
+                
+            }
+            else
+            {
+                json = File.ReadAllText("config.json", new UTF8Encoding(false));
             }
 
-            json = File.ReadAllText("../../config.json", new UTF8Encoding(false));
             cfg = JsonConvert.DeserializeObject<Config>(json);
 
             var tskl = new List<Task>();
